% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_marginals.R
\name{check_marginals}
\alias{check_marginals}
\title{Check that user-specified marginal distributions are supported}
\usage{
check_marginals(family, d)
}
\arguments{
\item{family}{List. A list of length d containing the names of marginal distributions}

\item{d}{Numeric. The number of variables}
}
\description{
Check that user-specified marginal distributions are supported
}
\keyword{internal}
