% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{comMat}
\alias{comMat}
\title{Commutation matrix}
\usage{
comMat(m, n = m)
}
\arguments{
\item{m}{Integer. First dimension.}

\item{n}{Integer. Second dimension (defaults to m).}
}
\value{
An mn x mn commutation matrix.
}
\description{
Constructs the mn x mn commutation matrix.
}
\keyword{internal}
