% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companion_form_phi.R
\name{companion_form_phi}
\alias{companion_form_phi}
\title{Construct companion form of VAR coefficient matrices}
\usage{
companion_form_phi(Phi, d, p)
}
\arguments{
\item{Phi}{Array. A d x d x p array of VAR coefficient matrices.}

\item{d}{Numeric. The number of variables.}

\item{p}{Numeric. The VAR order.}
}
\value{
A (p\emph{d) x (p}d) companion matrix.
}
\description{
Construct companion form of VAR coefficient matrices
}
\keyword{internal}
