% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{dupMat}
\alias{dupMat}
\title{Duplication matrix}
\usage{
dupMat(n, type = "s")
}
\arguments{
\item{n}{Integer. Matrix dimension.}

\item{type}{Character. Type of duplication: \code{"s"} (symmetric), \code{"l"}
(strictly lower triangular), or \code{"d"} (diagonal).}
}
\value{
A duplication matrix.
}
\description{
Duplication matrix
}
\keyword{internal}
