% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{eliMat}
\alias{eliMat}
\title{Elimination matrix}
\usage{
eliMat(n, type = "s")
}
\arguments{
\item{n}{Integer. Matrix dimension.}

\item{type}{Character. Type of elimination: \code{"s"} (symmetric), \code{"l"}
(strictly lower triangular), or \code{"d"} (diagonal).}
}
\value{
An elimination matrix.
}
\description{
Elimination matrix
}
\keyword{internal}
