% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_var.R
\name{fit_var}
\alias{fit_var}
\title{Get Yule-Walker estimates}
\usage{
fit_var(gamma_hat, Gamma_hat, d)
}
\arguments{
\item{gamma_hat}{Matrix. Stacked covariances matrix}

\item{Gamma_hat}{Matrix. Toeplitz covariance matrix}

\item{d}{Numeric. The number of variables}
}
\description{
Get Yule-Walker estimates
}
\keyword{internal}
