% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvar_resid_cov.R
\name{gvar_resid_cov}
\alias{gvar_resid_cov}
\title{Residual covariance for graphical VAR}
\usage{
gvar_resid_cov(A, S0_plus, S0_minus, S01, S10)
}
\arguments{
\item{A}{Matrix. d x d temporal coefficient matrix.}

\item{S0_plus}{Matrix. Lag-0 covariance of current-side variables.}

\item{S0_minus}{Matrix. Lag-0 covariance of past-side variables.}

\item{S01}{Matrix. Lag-1 past-current cross-covariance.}

\item{S10}{Matrix. Lag-1 current-past cross-covariance.}
}
\value{
A d x d symmetric residual covariance matrix.
}
\description{
Computes S_eps(A) = S0_plus - A\emph{S01 - S10}t(A) + A\emph{S0_minus}t(A).
}
\keyword{internal}
