% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation}
\alias{interpolation}
\title{Compute inverse link function to get latent covariances}
\usage{
interpolation(coef, u, v)
}
\arguments{
\item{coef}{Numeric. A vector of link function coefficients}

\item{u}{Numeric. A vector of grid values}

\item{v}{Numeric. A vector of observed covariances}
}
\description{
Compute inverse link function to get latent covariances
}
\keyword{internal}
