% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_var_invlink.R
\name{latent_var_invlink}
\alias{latent_var_invlink}
\title{Compute latent Gaussian covariances}
\usage{
latent_var_invlink(cov_x_hat, d, p, ell_ij_hat)
}
\arguments{
\item{cov_x_hat}{Array. An array of observed covariance matrices}

\item{d}{Numeric. Number of variables}

\item{p}{Numeric. The VAR order. Default is 1}

\item{ell_ij_hat}{Array. An array of link function coefficients}
}
\description{
Compute latent Gaussian covariances
}
\keyword{internal}
