% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_var_link.R
\name{latent_var_link}
\alias{latent_var_link}
\title{Latent link function}
\usage{
latent_var_link(data, d, n, k, family, ordinal_levels, corr)
}
\arguments{
\item{data}{Matrix. A d by n multivariate time series matrix}

\item{d}{Numeric. The number of variables}

\item{n}{Numeric. Time series length}

\item{k}{Numeric. The value at which Hermite coefficient infinite sums terminate. Default is 100}

\item{family}{List. A list of length d with the names of each distribution}

\item{ordinal_levels}{List. A list of length d containing the numeric levels
for each Ordinal variable and \code{NULL} for each non-Ordinal variable.}

\item{corr}{Logical. Correlations or covariances}
}
\value{
A k x d x d array of link function coefficients.
}
\description{
Latent link function
}
\keyword{internal}
