% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_var_link.R
\name{latent_var_link_numderiv}
\alias{latent_var_link_numderiv}
\title{Latent link function for numderiv}
\usage{
latent_var_link_numderiv(d, n, param_hat, family, ordinal_levels, corr)
}
\arguments{
\item{d}{Numeric. The number of variables}

\item{n}{Numeric. Time series length}

\item{param_hat}{List. List of estimated marginal parameters}

\item{family}{List. List of marginal distributions}

\item{ordinal_levels}{List. A list of length d containing the numeric levels
for each Ordinal variable and \code{NULL} for each non-Ordinal variable.}

\item{corr}{Logical. Correlations or covariances}
}
\value{
A k x d x d array of link function coefficients.
}
\description{
Latent link function for numderiv
}
\keyword{internal}
