% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{make_e_basis}
\alias{make_e_basis}
\title{Create unit basis vector e_i}
\usage{
make_e_basis(i, n)
}
\arguments{
\item{i}{Integer. Position of the 1.}

\item{n}{Integer. Length of the vector.}
}
\value{
A numeric vector of length n with 1 in position i.
}
\description{
Create unit basis vector e_i
}
\keyword{internal}
