% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{make_u_basis}
\alias{make_u_basis}
\title{Create symmetric unit basis vector}
\usage{
make_u_basis(i, j, n)
}
\arguments{
\item{i}{Integer. Row index.}

\item{j}{Integer. Column index.}

\item{n}{Integer. Matrix dimension.}
}
\value{
A numeric unit vector of length n(n+1)/2.
}
\description{
Create symmetric unit basis vector
}
\keyword{internal}
