% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{make_v_basis}
\alias{make_v_basis}
\title{Create strictly lower triangular unit basis vector}
\usage{
make_v_basis(i, j, n)
}
\arguments{
\item{i}{Integer. Row index.}

\item{j}{Integer. Column index.}

\item{n}{Integer. Matrix dimension.}
}
\value{
A numeric unit vector of length n(n-1)/2.
}
\description{
Create strictly lower triangular unit basis vector
}
\keyword{internal}
