% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{makeCommutationMat}
\alias{makeCommutationMat}
\title{Commutation matrix (transposed)}
\usage{
makeCommutationMat(n, m = n)
}
\arguments{
\item{n}{Integer. First dimension.}

\item{m}{Integer. Second dimension (defaults to n).}
}
\value{
A transposed commutation matrix.
}
\description{
Commutation matrix (transposed)
}
\keyword{internal}
