% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat_spline.R
\name{nat_spline}
\alias{nat_spline}
\title{Compute splines}
\usage{
nat_spline(coef, u)
}
\arguments{
\item{coef}{Numeric. A vector of link function coefficients}

\item{u}{Numeric. A vector of grid values}
}
\value{
A list with three components: spline coefficients matrix, second
derivatives vector, and interval widths vector.
}
\description{
Compute splines
}
\keyword{internal}
