% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{nMat}
\alias{nMat}
\title{Symmetrizer matrix N}
\usage{
nMat(n)
}
\arguments{
\item{n}{Integer. Matrix dimension.}
}
\value{
An n^2 x n^2 symmetrizer matrix.
}
\description{
Symmetrizer matrix N
}
\keyword{internal}
