% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.timecop_fit}
\alias{print.summary.timecop_fit}
\title{Print a timecop model summary}
\usage{
\method{print}{summary.timecop_fit}(x, ...)
}
\arguments{
\item{x}{A \code{summary.timecop_fit} object.}

\item{...}{Additional arguments passed to \code{\link[stats]{printCoefmat}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a timecop model summary
}
