% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.timecop_gvar}
\alias{print.summary.timecop_gvar}
\title{Print a graphical VAR model summary}
\usage{
\method{print}{summary.timecop_gvar}(x, ...)
}
\arguments{
\item{x}{A \code{summary.timecop_gvar} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a graphical VAR model summary
}
