% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.timecop_fit}
\alias{print.timecop_fit}
\title{Print a fitted timecop model}
\usage{
\method{print}{timecop_fit}(x, ...)
}
\arguments{
\item{x}{A \code{timecop_fit} object returned by \code{\link{fit_timecop}}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fitted timecop model
}
