% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.timecop_gvar}
\alias{print.timecop_gvar}
\title{Print a fitted graphical VAR model}
\usage{
\method{print}{timecop_gvar}(x, ...)
}
\arguments{
\item{x}{A \code{timecop_gvar} object returned by
\code{\link{fit_graphical_var}}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fitted graphical VAR model
}
