% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{revVech}
\alias{revVech}
\title{Reverse vech operator}
\usage{
revVech(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A symmetric matrix.
}
\description{
Reverse vech operator
}
\keyword{internal}
