% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{revVecp}
\alias{revVecp}
\title{Reverse vecp operator}
\usage{
revVecp(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A symmetric matrix with unit diagonal.
}
\description{
Reverse vecp operator
}
\keyword{internal}
