% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_var.R
\name{se_var}
\alias{se_var}
\title{Compute standard errors}
\usage{
se_var(
  data,
  gamma_hat,
  Gamma_hat,
  cov_x_hat,
  d,
  p,
  n,
  family,
  ordinal_levels,
  marg_num,
  corr
)
}
\arguments{
\item{data}{Matrix. A d by n multivariate time series matrix}

\item{gamma_hat}{Matrix. Stacked covariances matrix}

\item{Gamma_hat}{Matrix. Toeplitz covariance matrix}

\item{cov_x_hat}{Array. An array of observed covariances}

\item{d}{Numeric. The number of variables}

\item{p}{Numeric. The VAR order}

\item{n}{Numeric. Time series length}

\item{family}{List. A list of marginal distributions}

\item{ordinal_levels}{List. A list of length d containing the numeric levels
for each Ordinal variable and \code{NULL} for each non-Ordinal variable.}

\item{marg_num}{Numeric. The total number of marginal parameters}

\item{corr}{Logical. Correlations or covariances (SEs only available for covariances)}
}
\value{
A d x d matrix of standard errors for the VAR(1) coefficient estimates.
}
\description{
Compute standard errors
}
\keyword{internal}
