% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_data.R
\name{setup_data}
\alias{setup_data}
\title{Check and setup data}
\usage{
setup_data(data)
}
\arguments{
\item{data}{Matrix.  An n (time points) by d (variables) multivariate time series matrix}
}
\description{
Check and setup data
}
\keyword{internal}
