% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ordinal.R
\name{setup_ordinal_levels}
\alias{setup_ordinal_levels}
\title{Setup ordinal levels}
\usage{
setup_ordinal_levels(data, family, ordinal_levels, d)
}
\arguments{
\item{data}{Matrix. A d by n multivariate time series matrix}

\item{family}{List. A list of length d with the names of each distribution}

\item{ordinal_levels}{List. A list containing one numeric vector of levels for each ordinal variable}

\item{d}{Numeric. The number of variables}
}
\value{
A list of length d containing levels for each ordinal variable and NULL otherwise.
}
\description{
Setup ordinal levels
}
\keyword{internal}
