% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecopObjectClass.R
\docType{class}
\name{timecop-class}
\alias{timecop-class}
\title{timecop object class}
\description{
An object to be used with fit_timecop
}
\section{Slots}{

\describe{
\item{\code{data}}{Matrix. A d (variables) by n (time points) multivariate time series matrix}

\item{\code{cov_x_hat}}{Array. An array of observed covariances}

\item{\code{cov_z_hat}}{Array. An array of latent covariances}

\item{\code{gamma_hat}}{Matrix. A matrix of stacked covariances (lag 1 to p)}

\item{\code{Gamma_hat}}{Matrix. A Toeplitz matrix of covariances}

\item{\code{d}}{Numeric. The number of variables}

\item{\code{p}}{Numeric. The VAR order. Default is 1}

\item{\code{n}}{Numeric. Time series length}

\item{\code{marg_num}}{Numeric. Total number of marginal parameters}

\item{\code{family}}{List. A list of marginal distributions}

\item{\code{ordinal_levels}}{List. A list of length d containing the numeric levels
for each Ordinal variable and \code{NULL} for each non-Ordinal variable.}

\item{\code{pd_approx}}{Logical. A logical indicating if adjustments were made to ensure positive definiteness}

\item{\code{corr}}{Logical. A logical indicating if observed correlations or covariances were used}
}}

