% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecop-package.R
\docType{package}
\name{timecop-package}
\alias{timecop-package}
\title{timecop: Estimation of Latent Gaussian VAR Models for Discrete-Valued Time Series}
\description{
Estimates latent Gaussian vector autoregressive (VAR) models for discrete-valued multivariate time series. Supports Bernoulli, Poisson, and Gaussian marginal distributions through a latent variable approach using Hermite polynomial expansion. Estimation is performed via a method-of-moments Yule-Walker procedure with standard errors derived from the long-run variance of the estimating equations. Currently supports VAR order p=1.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crawfordcm/timecop}
  \item Report bugs at \url{https://github.com/crawfordcm/timecop/issues}
}

}
\author{
\strong{Maintainer}: Christopher Crawford \email{cmcrawford4792@gmail.com}

Authors:
\itemize{
  \item Christopher Crawford \email{cmcrawford4792@gmail.com}
  \item Younghoon Kim
}

Other contributors:
\itemize{
  \item Zachary Fisher \email{fish.zachary@gmail.com} [contributor]
  \item Vladas Pipiras [contributor]
  \item Duker Marie [contributor]
}

}
\keyword{internal}
