% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timecop_example}
\alias{timecop_example}
\title{Example bivariate Bernoulli time series}
\format{
A 200 x 2 numeric matrix. Each column is a binary (0/1) time series
of length 200.
}
\source{
Simulated data.
}
\usage{
timecop_example
}
\description{
A simulated 200 x 2 matrix of bivariate Bernoulli time series data generated
from a latent Gaussian VAR(1) process. Generated using \code{\link[=latent_var_sim]{latent_var_sim()}}
with \code{set.seed(42)}, success probabilities of 0.5 for both variables, and
a transition matrix with 0.4 on the diagonal and 0.2 on the off-diagonal.
}
\keyword{datasets}
