% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{vecd}
\alias{vecd}
\title{Diagonal vectorization}
\usage{
vecd(x)
}
\arguments{
\item{x}{A square matrix.}
}
\value{
A column vector of diagonal elements.
}
\description{
Diagonal vectorization
}
\keyword{internal}
