% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{vech}
\alias{vech}
\title{Half-vectorize a symmetric matrix}
\usage{
vech(x)
}
\arguments{
\item{x}{A symmetric matrix.}
}
\value{
A column vector of lower triangular elements including diagonal.
}
\description{
Half-vectorize a symmetric matrix
}
\keyword{internal}
