% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcalc.R
\name{vecp}
\alias{vecp}
\title{Strict lower triangular vectorization}
\usage{
vecp(x)
}
\arguments{
\item{x}{A square matrix.}
}
\value{
A column vector of strictly lower triangular elements.
}
\description{
Strict lower triangular vectorization
}
\keyword{internal}
