% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pd.R
\name{check_pd}
\alias{check_pd}
\title{Check if covariance matrices are PD}
\usage{
check_pd(gamma_hat, Gamma_hat, d, eig.tol)
}
\arguments{
\item{gamma_hat}{Matrix. Stacked covariance matrix}

\item{Gamma_hat}{Matrix. Toeplitz covariance matrix}

\item{d}{Numeric. The number of variables}

\item{eig.tol}{Numeric. Eigen tolerance}
}
\value{
A list of two matrices: the (possibly adjusted) \code{gamma_hat} and
\code{Gamma_hat}.
}
\description{
Check if covariance matrices are PD
}
\keyword{internal}
