% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_marginal_params.R
\name{estimate_marginal_params}
\alias{estimate_marginal_params}
\title{Estimate marginal parameters from data}
\usage{
estimate_marginal_params(data, family, d)
}
\arguments{
\item{data}{Matrix. A d by n multivariate time series matrix}

\item{family}{List. A list of length d with the names of each distribution}

\item{d}{Numeric. The number of variables}
}
\value{
A list of length d of estimated marginal parameters.
}
\description{
Estimate marginal parameters from data
}
\keyword{internal}
