% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_var_link_numderiv.R
\name{latent_var_link_numderiv}
\alias{latent_var_link_numderiv}
\title{Latent link function for numderiv}
\usage{
latent_var_link_numderiv(d, n, param_hat, family)
}
\arguments{
\item{d}{Numeric. The number of variables}

\item{n}{Numeric. Time series length}

\item{param_hat}{List. List of estimated marginal parameters}

\item{family}{List. List of marginal distributions}
}
\description{
Latent link function for numderiv
}
\keyword{internal}
