% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_var_sim.R
\name{latent_var_sim}
\alias{latent_var_sim}
\title{Simulate multivariate time series}
\usage{
latent_var_sim(d, n, p, param, phi_lv, family)
}
\arguments{
\item{d}{Numeric. The number of variables}

\item{n}{Numeric. Time series length}

\item{p}{Numeric. The VAR order}

\item{param}{List. A list of marginal parameters}

\item{phi_lv}{Matrix. A d by d transition matrix}

\item{family}{List. A list of marginal distributions}
}
\description{
Simulate multivariate time series
}
