% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longrun_var.R
\name{longrun_var}
\alias{longrun_var}
\title{Compute long-run variance}
\usage{
longrun_var(data, d, n, family)
}
\arguments{
\item{data}{Matrix. A d by n multivariate time series matrix}

\item{d}{Numeric. The number of variables}

\item{n}{Numeric. Time series length}

\item{family}{List. A list of marginal distributions}
}
\value{
A long-run variance-covariance matrix.
}
\description{
Compute long-run variance
}
\keyword{internal}
