% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numderiv.R
\name{numderiv}
\alias{numderiv}
\title{Compute numerical derivatives}
\usage{
numderiv(data, cov_x_hat, d, p, n, family, corr)
}
\arguments{
\item{data}{Matrix. A d by n multivariate time series matrix}

\item{cov_x_hat}{Array. An array of observed covariances}

\item{d}{Numeric. The number of variables}

\item{p}{Numeric. The VAR order}

\item{n}{Numeric. Time series length}

\item{family}{List. A list of marginal distributions}

\item{corr}{Logical. Correlations or covariances}
}
\value{
A Jacobian matrix of numerical derivatives.
}
\description{
Compute numerical derivatives
}
\keyword{internal}
