% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_var_cov.R
\name{observed_var_cov}
\alias{observed_var_cov}
\title{Compute observed covariances}
\usage{
observed_var_cov(data, d, p, n)
}
\arguments{
\item{data}{Matrix. An n (time points) by d (variables) multivariate time series matrix}

\item{d}{Numeric. The number of variables}

\item{p}{Numeric. The VAR order. Default is 1}

\item{n}{Numeric. Time series length}
}
\description{
Compute observed covariances
}
\keyword{internal}
