% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecopObjectClass.R
\name{timecop}
\alias{timecop}
\title{Construct an object of class timecop}
\usage{
timecop(data = NULL, family = NULL, p = 1)
}
\arguments{
\item{data}{Matrix. An n (time points) by d (variables) multivariate time series matrix}

\item{family}{List. A list of length d with the names of each distribution}

\item{p}{Numeric. The VAR order. Default is 1}
}
\description{
Construct an object of class timecop
}
